Extract BotManager protocol, shared formatting, and rate limiter
Pure refactor — no behavior changes. Extracts common bot infrastructure
from telegram.py into reusable modules for the Discord adapter (#71).

- bot_manager.py: BotAdapter typing.Protocol (send_notification, handle_event, start, stop)
- bot_formatting.py: Platform-agnostic plain-text formatters and data access
- bot_ratelimit.py: RateLimiter utility class (same-cmd 5s, cross-cmd 1s)
- telegram.py: Imports from shared modules, removes duplicated code
- Test imports updated to new module locations

Ref #73
